refactor(worker): the research job keeps the job, not the research
execute-personalization-research.ts was 770 lines. The job handler and the run
loop were 140 of them; the rest was everything the run needs done: reading a
homepage, searching, calling Apollo, prompting for a profile and cards, then
tidying what came back.

Each of those is its own file now, and the job file reads as the sequence it
runs. Where a constant or schema had one user it went with it; the few that are
genuinely shared sit in personalization-research-shared, which imports nothing.

Nothing is rewritten, and the imports run one way: shaping reads shared,
enrichment and generate read shaping, sources stands alone, the job reads all
of them. Normalise away the imports and the new files match the original line
for line, in both directions.
